requirements.txt
vit-pytorch
performer-pytorch
pytorch-fast-transformers>=0.3.0
nystrom-attention
tensorboard